requirements.txt
flask
yt-dlp
celery
redis
gunicorn
gevent
customtkinter
Pillow
requests